Move get_substring_geometry from utils to a classmethod of LinearIntegrator
