+++ tests/test_importers/test_utils.py
@@ -15,8 +15,7 @@ from qgis.core import (
     QgsPointXY,
 )
 
-from threedi_schematisation_editor.custom_tools.utils import update_attributes, FeatureManager, get_substring_geometry, \
-    ColumnImportMethod
+from threedi_schematisation_editor.custom_tools.utils import update_attributes, FeatureManager, ColumnImportMethod
 
 
 @pytest.fixture
@@ -138,37 +137,6 @@ def test_update_attributes_type_conversion_error():
 
     # Assert
     assert new_feat['id'] == NULL
-
-
-@pytest.mark.parametrize('simplify', [True, False])
-def test_get_substring_geometry_argument_processing(simplify):
-    """Test that get_substring_geometry processes the simplify argument correctly."""
-    # Create a simple line geometry for testing
-    line_geom = QgsGeometry.fromPolylineXY([
-        QgsPointXY(0, 0),
-        QgsPointXY(50, 0),
-        QgsPointXY(100, 0)
-    ])
-
-    # Get the underlying curve object
-    curve = line_geom.constGet()
-
-    # Test parameters
-    start_distance = 25.0
-    end_distance = 75.0
-
-    # Call the function with simplify=False
-    result = get_substring_geometry(curve, start_distance, end_distance, simplify=simplify)
-
-    # Verify that both calls return a QgsGeometry object
-    assert isinstance(result, QgsGeometry)
-
-    if simplify:
-        assert len(result.asPolyline()) == 2
-    else:
-        assert len(result.asPolyline()) >= 2
-
-    assert result.length() == end_distance - start_distance
 
 
 @pytest.fixture
